The subsystem seems to work =) We can start a client in a subsystem and let him use a service living in the root-subsystem and let them share a dataspace.

+++ include/pthread.h
@@ -21,7 +21,7 @@
 #include <arch/Types.h>
 #include <arch/SpinLock.h>
 
-typedef size_t pthread_key_t;
+typedef uint32_t pthread_key_t;
 typedef spinlock_t pthread_mutex_t;
 typedef int pthread_t;
 typedef int pthread_once_t;
